requirements.txt
beautifulsoup4==4.10.0
suricata-update==1.3.0

